"Ajout des fonctionnalités SOAP Server de reservation."

+++ biblio-ws/biblio-ws-business/src/main/java/oc/projet/biblio/business/service/OuvrageService.java
@@ -23,5 +23,5 @@ public interface OuvrageService {
 
     LocalDate firstDispoDate(Ouvrage ouvrage);
 
-    Long countResa(Ouvrage ouvrage);
+    int countResa(Ouvrage ouvrage);
 }
